Create writer file only if metrics are present (#23)
* Create writer file only if metrics are present


* Add metrics to writers metric

* Rename writer_file_creation metric name


* Update the statistic for the metric

* Spotless changes

+++ src/main/java/org/opensearch/performanceanalyzer/reader_writer_shared/EventLogFileHandler.java
@@ -82,8 +82,11 @@ public class EventLogFileHandler {
      * data.
      *
      * <p>If any of the above steps fail, then the tmp file is not deleted from the filesystem. This
-     * is fine as the {@link org.opensearch.performanceanalyzer.reader_writer_shared.EventLogFileHandler#deleteFiles},
-     * will eventually clean it. The copies are atomic and therefore the reader never reads incompletely written file.
+     * is fine as the {@link
+     * org.opensearch.performanceanalyzer.reader_writer_shared.EventLogFileHandler#deleteFiles},
+     * will eventually clean it. The copies are atomic and therefore the reader never reads
+     * incompletely written file.
+     *
      * @param dataEntries The metrics to be written to file.
      * @param epoch The epoch all the metrics belong to.
      */
@@ -116,11 +119,19 @@ public class EventLogFileHandler {
     public void renameFromTmpWithPrivilege(long epoch) {
         Path path = Paths.get(metricsLocation, String.valueOf(epoch));
         Path tmpPath = Paths.get(path.toString() + TMP_FILE_EXT);
-        // This is done only when no exception is thrown.
-        try {
-            Files.move(tmpPath, path, REPLACE_EXISTING, ATOMIC_MOVE);
-        } catch (IOException e) {
-            LOG.error("Error moving file {} to {}.", tmpPath.toString(), path.toString(), e);
+        File tempFile = new File(tmpPath.toString());
+        // Only if the tmp file is present, we want to create the writer file.
+        // If not, we will publish a metric.
+        if (tempFile.exists()) {
+            // This is done only when no exception is thrown.
+            try {
+                Files.move(tmpPath, path, REPLACE_EXISTING, ATOMIC_MOVE);
+            } catch (IOException e) {
+                LOG.error("Error moving file {} to {}.", tmpPath.toString(), path.toString(), e);
+            }
+        } else {
+            PerformanceAnalyzerApp.WRITER_METRICS_AGGREGATOR.updateStat(
+                    WriterMetrics.WRITER_FILE_CREATION_SKIPPED, "", 1);
         }
     }
 
@@ -134,7 +145,7 @@ public class EventLogFileHandler {
         File tempFile = new File(pathToFile.toString());
         if (!tempFile.exists()) {
             long mCurrT = System.currentTimeMillis();
-            LOG.info("Didnt find {} at {}", filename, mCurrT);
+            LOG.debug("Didnt find {} at {}", filename, mCurrT);
             return;
         }
         readInternal(pathToFile, BUFFER_SIZE, processor);
